requirements.txt
django==4.2.21
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
pillow==12.2.0
django-filter==24.3
gunicorn==25.3.0
python-decouple==3.8
whitenoise==6.12.0
